requirements.txt
discord.py>=2.5.2
Unidecode>=1.4.0
aiosqlite>=0.21.0
python-dotenv>=1.1.1
aiohttp>=3.12.13
structlog>=24.1.0
